Feat: started replacing old tipmanager
