+++ pkg/protocol/tipmanager/tip_metadata.go
@@ -1,0 +1,4 @@
+package tipmanager
+
+type TipMetadata interface {
+}
